refactor(kdv_pinn): clean up dead-code markers and unwrap usage
