+++ kdv_pinn/src/network.rs
@@ -2,13 +2,11 @@ use tenferro_runtime::{DType, DotGeneralConfig, TracedTensor};
 use tenferro_tensor::Tensor;
 
 /// A fully-connected layer using `TracedTensor` placeholders for weight and bias.
-#[allow(dead_code)]
 pub(crate) struct Linear {
     pub weight: TracedTensor,
     pub bias: TracedTensor,
 }
 
-#[allow(dead_code)]
 impl Linear {
     /// Create a new `Linear` layer with the given input/output feature sizes.
     pub fn new(in_features: usize, out_features: usize) -> Self {
@@ -39,12 +37,10 @@ impl Linear {
 }
 
 /// A multi-layer perceptron built from `Linear` layers with `tanh` activations.
-#[allow(dead_code)]
 pub(crate) struct Mlp {
     layers: Vec<Linear>,
 }
 
-#[allow(dead_code)]
 impl Mlp {
     /// Create a new `Mlp` from a slice of layer sizes.
     pub fn new(layer_sizes: &[usize]) -> Self {
@@ -88,7 +84,10 @@ impl Mlp {
         use rand::distributions::{Distribution, Uniform};
         let mut tensors = Vec::new();
         for layer in &self.layers {
-            let shape = layer.weight.try_concrete_shape().unwrap();
+            let shape = layer
+                .weight
+                .try_concrete_shape()
+                .expect("weight placeholder must have a concrete shape");
             let fan_in = shape[0];
             let fan_out = shape[1];
             let scale = (6.0 / (fan_in + fan_out) as f64).sqrt();
@@ -98,7 +97,10 @@ impl Mlp {
                 .collect();
             tensors.push(Tensor::from_vec_col_major(shape.clone(), w_data));
 
-            let b_shape = layer.bias.try_concrete_shape().unwrap();
+            let b_shape = layer
+                .bias
+                .try_concrete_shape()
+                .expect("bias placeholder must have a concrete shape");
             let b_data = vec![0.0_f64; b_shape.iter().product::<usize>()];
             tensors.push(Tensor::from_vec_col_major(b_shape.clone(), b_data));
         }
@@ -122,48 +124,4 @@ impl Mlp {
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use tenferro_cpu::CpuBackend;
-    use tenferro_runtime::{GraphCompiler, GraphExecutor, TracedTensor};
-    use tenferro_tensor::Tensor;
-
-    #[test]
-    fn mlp_forward_shape() {
-        let net = Mlp::new(&[2, 16, 16, 1]);
-        let x = TracedTensor::input_concrete_shape(DType::F64, &[4, 2]);
-        let y = net.forward(&x);
-        assert_eq!(y.rank, 2);
-        assert_eq!(y.try_concrete_shape(), Some(vec![4, 1]));
-
-        let specs = net.input_specs();
-        let bindings: Vec<(&TracedTensor, DType, &[usize])> = specs
-            .iter()
-            .map(|(p, dtype, shape)| (*p, *dtype, shape.as_slice()))
-            .chain(std::iter::once((&x, DType::F64, &[4, 2][..])))
-            .collect();
-
-        let mut compiler = GraphCompiler::new();
-        let program = compiler.compile_with_input_specs(&y, &bindings).unwrap();
-        let mut executor = GraphExecutor::new(CpuBackend::new());
-
-        let x_tensor = Tensor::from_vec_col_major(vec![4, 2], vec![0.0_f64; 8]);
-        let param_tensors: Vec<Tensor> = specs
-            .iter()
-            .map(|(_, _, shape)| {
-                let len = shape.iter().product::<usize>();
-                Tensor::from_vec_col_major(shape.clone(), vec![0.1_f64; len])
-            })
-            .collect();
-        let mut bind_tensors: Vec<(&TracedTensor, &Tensor)> = net
-            .parameters()
-            .iter()
-            .zip(param_tensors.iter())
-            .map(|(p, t)| (*p, t))
-            .collect();
-        bind_tensors.push((&x, &x_tensor));
-
-        let out = executor.run_with_inputs(&program, &bind_tensors).unwrap();
-        assert_eq!(out.shape(), &[4, 1]);
-    }
-}
+mod tests;
