+++ kdv_pinn/src/sampler.rs
@@ -1,6 +1,3 @@
-// TODO: remove once Sampler is wired into the training loop / main.
-#![allow(dead_code)]
-
 use rand::distributions::{Distribution, Uniform};
 use rand::Rng;
 use tenferro_tensor::Tensor;
@@ -8,7 +5,10 @@ use tenferro_tensor::Tensor;
 pub struct Sampler {
     x_min: f64,
     x_max: f64,
+    // TODO(kdv-pinn): remove #[allow(dead_code)] once Tasks 9-10 wire t_min/t_max.
+    #[allow(dead_code)]
     t_min: f64,
+    #[allow(dead_code)]
     t_max: f64,
 }
 
@@ -22,6 +22,8 @@ impl Sampler {
         }
     }
 
+    // TODO(kdv-pinn): remove #[allow(dead_code)] once Tasks 9-10 wire collocation.
+    #[allow(dead_code)]
     pub fn collocation<R: Rng>(&self, n: usize, rng: &mut R) -> Tensor {
         let x_dist = Uniform::new(self.x_min, self.x_max);
         let t_dist = Uniform::new(self.t_min, self.t_max);
@@ -48,6 +50,8 @@ impl Sampler {
         )
     }
 
+    // TODO(kdv-pinn): remove #[allow(dead_code)] once Tasks 9-10 wire boundary.
+    #[allow(dead_code)]
     pub fn boundary<R: Rng>(&self, n: usize, rng: &mut R) -> (Tensor, Tensor, Tensor) {
         let t_dist = Uniform::new(self.t_min, self.t_max);
         let mut x = Vec::with_capacity(n);
